+++ realm/realm-library/src/syncIntegrationTest/java/io/realm/SyncedRealmTests.java
@@ -17,7 +17,6 @@
 package io.realm;
 
 import android.os.SystemClock;
-import android.support.annotation.NonNull;
 import android.support.test.annotation.UiThreadTest;
 import android.support.test.runner.AndroidJUnit4;
 
@@ -27,15 +26,14 @@ import org.junit.runner.RunWith;
 import java.io.File;
 import java.util.Random;
 import java.util.UUID;
-import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicReference;
 
 import io.realm.entities.StringOnly;
 import io.realm.exceptions.DownloadingRealmInterruptedException;
 import io.realm.exceptions.RealmMigrationNeededException;
 import io.realm.objectserver.utils.Constants;
 import io.realm.rule.RunTestInLooperThread;
+import io.realm.util.SyncTestUtils;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -51,8 +49,7 @@ public class SyncedRealmTests extends BaseIntegrationTest {
     @Test
     @UiThreadTest
     public void waitForInitialRemoteData_mainThreadThrows() {
-        final SyncUser user = loginUser();
-
+        final SyncUser user = SyncTestUtils.createTestUser(Constants.AUTH_URL);
         SyncConfiguration config = new SyncConfiguration.Builder(user, Constants.USER_REALM)
                 .waitForInitialRemoteData()
                 .build();
@@ -67,23 +64,6 @@ public class SyncedRealmTests extends BaseIntegrationTest {
                 realm.close();
             }
         }
-    }
-
-    // Login user on a worker thread, so this method can be used from both UI and non-ui threads.
-    @NonNull
-    private SyncUser loginUser() {
-        final CountDownLatch userReady = new CountDownLatch(1);
-        final AtomicReference<SyncUser> user = new AtomicReference<>();
-        new Thread(new Runnable() {
-            @Override
-            public void run() {
-                SyncCredentials credentials = SyncCredentials.usernamePassword(UUID.randomUUID().toString(), "password", true);
-                user.set(SyncUser.login(credentials, Constants.AUTH_URL));
-                userReady.countDown();
-            }
-        }).start();
-        TestHelper.awaitOrFail(userReady);
-        return user.get();
     }
 
     @Test